[#244] Refactor fsw
